Semana13/1 - Demo StreamLit/requirements.txt
opencv_python==4.5.4.60
streamlit==1.3.0
mediapipe==0.8.9.1
numpy==1.19.3
Pillow==8.2.0
altair<5